OSD attributes updated - MTextOSD has 10 of stroke width by default. - some bug fixed
git-svn-id: http://movist.googlecode.com/svn/trunk@75 9988c26d-9134-0410-b5bb-5778289bb252

+++ PreferenceController_Subtitle.m
@@ -60,8 +60,8 @@
     [_subtitleShadowBlurTextField setFloatValue:shadowBlur];
     [_subtitleShadowOffsetSlider setFloatValue:shadowOffset];
     [_subtitleShadowOffsetTextField setFloatValue:shadowOffset];
-    [_subtitleShadowDarknessSlider setFloatValue:shadowDarkness];
-    [_subtitleShadowDarknessTextField setFloatValue:shadowDarkness];
+    [_subtitleShadowDarknessSlider setIntValue:shadowDarkness];
+    [_subtitleShadowDarknessTextField setIntValue:shadowDarkness];
 
     [_subtitleDisplayOnLetterBoxButton setState:[_defaults boolForKey:MSubtitleDisplayOnLetterBoxKey]];
